+++ src/core/error/index.ts
@@ -1,14 +1,4 @@
-export class FractalSDKError extends Error {
-  name: string;
-  constructor(message: string) {
-    super(message);
-    this.name = 'FractalSDKError';
-
-    // 👇️ because we are extending a built-in class
-    Object.setPrototypeOf(this, FractalSDKError.prototype);
-  }
-
-  getUserFacingErrorMessage() {
-    return 'Something went wrong: ' + this.message;
-  }
-}
+export * from 'core/error/base';
+export * from 'core/error/approve';
+export * from 'core/error/auth';
+export * from 'core/error/transaction';
